refactor: Use Promise instead of Observable for calendar events
The google-calendar.service was using an Observable to return the result of a single HTTP call. This is unnecessary, as a Promise is a better fit for a single asynchronous value.

Refactored getPublicCalendarEvents to be an async function that returns a Promise<CalendarEvent[]>.

Updated EventListComponent to correctly handle the Promise returned by the service.